+++ src/Resources/contao/dca/tl_module.php
@@ -13,91 +13,114 @@
  * Add palette to tl_module
  */
 
-$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_einstellungen'] = '{title_legend},name,headline,type;{protected_legend:hide},protected;{expert_legend:hide},cssID,align,space';
-$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_spielersuche'] = '{title_legend},name,headline,type;{config_legend},dewis_searchfield;{protected_legend:hide},protected;{expert_legend:hide},cssID,align,space';
-$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_spieler'] = '{title_legend},name,headline,type;{protected_legend:hide},protected;{expert_legend:hide},cssID,align,space';
-$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_verein'] = '{title_legend},name,headline,type;{protected_legend:hide},protected;{expert_legend:hide},cssID,align,space';
-$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_verband'] = '{title_legend},name,headline,type;{protected_legend:hide},protected;{expert_legend:hide},cssID,align,space';
-$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_turnier'] = '{title_legend},name,headline,type;{protected_legend:hide},protected;{expert_legend:hide},cssID,align,space';
-$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_suche'] = '{title_legend},name,headline,type;{options_legend},dewis_searchkey,dewis_spieler,dewis_verein,dewis_verband,dewis_turnier;{protected_legend:hide},protected;{expert_legend:hide},cssID,align,space';
+$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_einstellungen'] = '{title_legend},name,headline,type;{protected_legend:hide},protected;{expert_legend:hide},cssID,align';
+$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_spielersuche'] = '{title_legend},name,headline,type;{config_legend},dewis_searchfield;{protected_legend:hide},protected;{expert_legend:hide},cssID,align';
+$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_spieler'] = '{title_legend},name,headline,type;{protected_legend:hide},protected;{expert_legend:hide},cssID,align';
+$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_verein'] = '{title_legend},name,headline,type;{protected_legend:hide},protected;{expert_legend:hide},cssID,align';
+$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_verband'] = '{title_legend},name,headline,type;{protected_legend:hide},protected;{expert_legend:hide},cssID,align';
+$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_turnier'] = '{title_legend},name,headline,type;{protected_legend:hide},protected;{expert_legend:hide},cssID,align';
+$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_suche'] = '{title_legend},name,headline,type;{options_legend},dewis_searchkey,dewis_spieler,dewis_verein,dewis_verband,dewis_turnier;{protected_legend:hide},protected;{expert_legend:hide},cssID,align';
+$GLOBALS['TL_DCA']['tl_module']['palettes']['dewis_bestenliste'] = '{title_legend},name,headline,type;{dwzbestenliste_legend},dwz_topcount,dwz_gender;{protected_legend:hide},protected;{expert_legend:hide},cssID,align';
 
 
 $GLOBALS['TL_DCA']['tl_module']['fields']['dewis_searchfield'] = array
 (
-	'label'                   => &$GLOBALS['TL_LANG']['tl_module']['dewis_searchfield'],
-	'inputType'               => 'checkbox',
-	'eval'                    => array
+	'label'                              => &$GLOBALS['TL_LANG']['tl_module']['dewis_searchfield'],
+	'inputType'                          => 'checkbox',
+	'eval'                               => array
 	(
-		'tl_class'            => 'w50',
-		'isBoolean'           => true,
+		'tl_class'                       => 'w50',
+		'isBoolean'                      => true,
 	),
-	'sql'                     => "char(1) NOT NULL default ''",
+	'sql'                                => "char(1) NOT NULL default ''",
 );
 
-// Welcher Parameter enthält den Suchstring?
+// Welcher Parameter enthÃ¤lt den Suchstring?
 $GLOBALS['TL_DCA']['tl_module']['fields']['dewis_searchkey'] = array
 (
-	'label'                   => &$GLOBALS['TL_LANG']['tl_module']['dewis_searchkey'],
-	'exclude'                 => true,
-	'default'                 => 'q',
-	'inputType'               => 'text',
-	'eval'                    => array
+	'label'                              => &$GLOBALS['TL_LANG']['tl_module']['dewis_searchkey'],
+	'exclude'                            => true,
+	'default'                            => 'q',
+	'inputType'                          => 'text',
+	'eval'                               => array
 	(
-		'mandatory'           => false,
-		'maxlength'           => 16,
-		'tl_class'            => 'w50'
+		'mandatory'                      => false,
+		'maxlength'                      => 16,
+		'tl_class'                       => 'w50'
 	),
-	'sql'                     => "varchar(16) NOT NULL default 'q'"
+	'sql'                                => "varchar(16) NOT NULL default 'q'"
 );
 
 // Suche in Spieler aktivieren/deaktivieren
 $GLOBALS['TL_DCA']['tl_module']['fields']['dewis_spieler'] = array
 (
-	'label'                   => &$GLOBALS['TL_LANG']['tl_module']['dewis_spieler'],
-	'inputType'               => 'checkbox',
-	'eval'                    => array
+	'label'                              => &$GLOBALS['TL_LANG']['tl_module']['dewis_spieler'],
+	'inputType'                          => 'checkbox',
+	'eval'                               => array
 	(
-		'tl_class'            => 'w50 clr',
-		'isBoolean'           => true,
+		'tl_class'                       => 'w50 clr',
+		'isBoolean'                      => true,
 	),
-	'sql'                     => "char(1) NOT NULL default ''",
+	'sql'                                => "char(1) NOT NULL default ''",
 );
 
 // Suche in Verein aktivieren/deaktivieren
 $GLOBALS['TL_DCA']['tl_module']['fields']['dewis_verein'] = array
 (
-	'label'                   => &$GLOBALS['TL_LANG']['tl_module']['dewis_verein'],
-	'inputType'               => 'checkbox',
-	'eval'                    => array
+	'label'                              => &$GLOBALS['TL_LANG']['tl_module']['dewis_verein'],
+	'inputType'                          => 'checkbox',
+	'eval'                               => array
 	(
-		'tl_class'            => 'w50',
-		'isBoolean'           => true,
+		'tl_class'                       => 'w50',
+		'isBoolean'                      => true,
 	),
-	'sql'                     => "char(1) NOT NULL default ''",
+	'sql'                                => "char(1) NOT NULL default ''",
 );
 
 // Suche in Verband aktivieren/deaktivieren
 $GLOBALS['TL_DCA']['tl_module']['fields']['dewis_verband'] = array
 (
-	'label'                   => &$GLOBALS['TL_LANG']['tl_module']['dewis_verband'],
-	'inputType'               => 'checkbox',
-	'eval'                    => array
+	'label'                              => &$GLOBALS['TL_LANG']['tl_module']['dewis_verband'],
+	'inputType'                          => 'checkbox',
+	'eval'                               => array
 	(
-		'tl_class'            => 'w50',
-		'isBoolean'           => true,
+		'tl_class'                       => 'w50',
+		'isBoolean'                      => true,
 	),
-	'sql'                     => "char(1) NOT NULL default ''",
+	'sql'                                => "char(1) NOT NULL default ''",
 );
 
 // Suche in Turnier aktivieren/deaktivieren
 $GLOBALS['TL_DCA']['tl_module']['fields']['dewis_turnier'] = array
 (
-	'label'                   => &$GLOBALS['TL_LANG']['tl_module']['dewis_turnier'],
-	'inputType'               => 'checkbox',
-	'eval'                    => array
+	'label'                              => &$GLOBALS['TL_LANG']['tl_module']['dewis_turnier'],
+	'inputType'                          => 'checkbox',
+	'eval'                               => array
 	(
-		'tl_class'            => 'w50',
-		'isBoolean'           => true,
+		'tl_class'                       => 'w50',
+		'isBoolean'                      => true,
 	),
-	'sql'                     => "char(1) NOT NULL default ''",
+	'sql'                                => "char(1) NOT NULL default ''",
 );
+
+$GLOBALS['TL_DCA']['tl_module']['fields']['dwz_topcount'] = array
+(
+	'label'                              => &$GLOBALS['TL_LANG']['tl_module']['dwz_topcount'],
+	'default'                            => 30,
+	'exclude'                            => true,
+	'inputType'                          => 'text',
+	'eval'                               => array('tl_class'=>'w50', 'rgxp'=>'digit', 'maxlength'=>6),
+	'sql'                                => "varchar(6) NOT NULL default ''"
+);
+
+$GLOBALS['TL_DCA']['tl_module']['fields']['dwz_gender'] = array
+(
+	'label'                              => &$GLOBALS['TL_LANG']['tl_module']['dwz_gender'],
+	'exclude'                            => true,
+	'default'                            => 'M',
+	'inputType'                          => 'select',
+	'options'                            => $GLOBALS['TL_LANG']['tl_module']['dwz_gender_options'],
+	'eval'                               => array('tl_class'=>'w50'),
+	'sql'                                => "char(1) NOT NULL default 'M'"
+);
+