+++ src/Resources/contao/languages/de/tl_module.php
@@ -17,3 +17,11 @@ $GLOBALS['TL_LANG']['tl_module']['dewis_spieler'] = array('Spielersuche','Suche 
 $GLOBALS['TL_LANG']['tl_module']['dewis_verein'] = array('Vereinssuche','Suche nach Vereinen aktivieren');
 $GLOBALS['TL_LANG']['tl_module']['dewis_verband'] = array('Verbandssuche','Suche nach Verbänden aktivieren');
 $GLOBALS['TL_LANG']['tl_module']['dewis_turnier'] = array('Turniersuche','Suche nach Turnieren aktivieren');
+$GLOBALS['TL_LANG']['tl_module']['dwz_topcount'] = array('Anzahl', 'Anzahl Spieler festlegen');
+$GLOBALS['TL_LANG']['tl_module']['dwz_gender'] = array('Geschlecht', 'Geschlecht für Liste festlegen. Standard ist männlich, also alle Spieler.');
+
+$GLOBALS['TL_LANG']['tl_module']['dwz_gender_options'] = array
+(
+	'm' => 'Alle Spieler',
+	'f' => 'Nur Frauen',
+);